Rename: Completion.java => RemoteCompletion.java
